Create separate TypeTraits headers for Reducer object and Kernel.
