+++ include/RAJA/pattern/detail/TypeTraits.hpp
@@ -1,3 +1,20 @@
+/*!
+ ******************************************************************************
+ *
+ * \file
+ *
+ * \brief   Header containing helper type traits for work with Reducers
+ *
+ ******************************************************************************
+ */
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+// Copyright (c) 2016-25, Lawrence Livermore National Security, LLC
+// and RAJA project contributors. See the RAJA/LICENSE file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+
 
 #ifndef RAJA_TYPETRAITS_HPP
 #define RAJA_TYPETRAITS_HPP
